Combine AzureWizardUserInput and AzureUserInput
