+++ ui/package.json
@@ -1,7 +1,7 @@
 {
     "name": "vscode-azureextensionui",
     "author": "Microsoft Corporation",
-    "version": "0.44.3",
+    "version": "0.45.0",
     "description": "Common UI tools for developing Azure extensions for VS Code",
     "tags": [
         "azure",
